Separate DSL tests into new test suite
